Hook GNOME Shell search provider up to tab hostnames
Search windows and tabs for a tab with a hostname that matches the
search string, and switch to that tab and window when the search
result is activated.

Currently, when multiple tabs match, only one search result is shown
and a random tab is selected; it's probably best to switch to the
last-active tab for that hostname rather than a random one.

https://bugzilla.gnome.org/show_bug.cgi?id=712345